lot click event; refactor
